Refactor: Split IdentifierPool.
